% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanDict}
\alias{cleanDict}
\title{Make a dictionary tibble from a keys vector and a values vector}
\usage{
cleanDict(keys, values)
}
\arguments{
\item{keys}{dictionary keys}

\item{values}{dictionary values}
}
\description{
Make a dictionary tibble from a keys vector and a values vector
}
