% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeDict}
\alias{mergeDict}
\title{merge a data vector with a dictionary table}
\usage{
mergeDict(dat, df = NULL, keys, values)
}
\arguments{
\item{dat}{data vector}

\item{df}{dictionary table}

\item{keys}{dictionary keys}

\item{values}{dictionary values}
}
\description{
merge a data vector with a dictionary table
}
