% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{catDict}
\alias{catDict}
\title{use mergeDict to apply a lookup table}
\usage{
catDict(dat, df = NULL, keys, values)
}
\arguments{
\item{dat}{data vector}

\item{df}{dictionary table}

\item{keys}{dictionary keys}

\item{values}{dictionary values}
}
\description{
use mergeDict to apply a lookup table
}
