% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{patchDict}
\alias{patchDict}
\title{use mergeDict to patch errors identified in a table}
\usage{
patchDict(dat, dict = NULL, keys, values)
}
\arguments{
\item{dat}{data vector}

\item{dict}{dictionary table}

\item{keys}{dictionary keys}

\item{values}{dictionary values}
}
\description{
use mergeDict to patch errors identified in a table
}
